+++ docs/CHANGES.md
@@ -2,6 +2,7 @@
 
 ## 3.0.1 (Unreleased)
 
+- #806 The `_converse.listen` API event listeners aren't triggered. [jcbrand]
 - #807 Error: Plugin "converse-dragresize" tried to override HeadlinesBoxView but it's not found. [jcbrand]
 
 ## 3.0.0 (2017-03-05)