+++ docs/CHANGES.md
@@ -4,6 +4,7 @@
 
 - #806 The `_converse.listen` API event listeners aren't triggered. [jcbrand]
 - #807 Error: Plugin "converse-dragresize" tried to override HeadlinesBoxView but it's not found. [jcbrand]
+- #628 Fixes the bug in displaying chat status during private chat (Message carbons) [saganshul]
 
 ## 3.0.0 (2017-03-05)
 
@@ -72,7 +73,6 @@
 - #585 Fixes the duplication bug due to case sensivity in adding contacts [saganshul]
 - #610, #785 Add presence priority handling [w3host, jcbrand]
 - #620 `auto_away` shouldn't change the user's status if it's set to `dnd`. [jcbrand]
-- #628 Fixes the bug in displaying chat status during private chat (Message carbons) [saganshul]
 - #694 The `notification_option` wasn't being used consistently. [jcbrand]
 - #745 New config option [priority](https://conversejs.org/docs/html/configuration.html#priority) [jcbrand]
 - #770 Allow setting contact attrs on chats.open [Ape]