+++ docs/CHANGES.md
@@ -4,7 +4,7 @@
 
 - #806 The `_converse.listen` API event listeners aren't triggered. [jcbrand]
 - #807 Error: Plugin "converse-dragresize" tried to override HeadlinesBoxView but it's not found. [jcbrand]
-- #628 Fixes the bug in displaying chat status during private chat (Message carbons) [saganshul]
+- #628 Fixes the bug in displaying chat status during private chat [saganshul]
 
 ## 3.0.0 (2017-03-05)