+++ docs/CHANGES.md
@@ -2,9 +2,9 @@
 
 ## 3.0.1 (Unreleased)
 
+- #628 Fixes the bug in displaying chat status during private chat. [saganshul]
 - #806 The `_converse.listen` API event listeners aren't triggered. [jcbrand]
 - #807 Error: Plugin "converse-dragresize" tried to override HeadlinesBoxView but it's not found. [jcbrand]
-- #628 Fixes the bug in displaying chat status during private chat [saganshul]
 
 ## 3.0.0 (2017-03-05)